+++ .talismanrc
@@ -410,7 +410,7 @@ fileignoreconfig:
 - filename: packages/shared-ui/src/composables/index.js
   checksum: 1a41578cc15aa2da2b59efa2c378e9ed2552a2912c7d7cd666527e148d4d15f6
 - filename: packages/shared-ui/src/composables/useAuthentication.ts
-  checksum: 685d0e302f58f682353b6b641108f4d7768743517fe7c58c30640d1ed096c3b3
+  checksum: 70d12f131247372acba4d2f5788499a3a7185ae6f4bd75cbf18ed4072536637f
 - filename: packages/shared-ui/src/composables/useToaster.ts
   checksum: 704ad956e4817e92ca2df2f4cbdf14cc7a39218cdacdfef070dd91c8d1e4c563
 - filename: packages/shared-ui/src/types/Auth.type.ts
